patch: move parsing related functions to lexer
