+++ src/parser/lexer.rs
@@ -1,6 +1,16 @@
-use std::sync::mpsc::Sender;
+use std::sync::mpsc::{channel, Receiver, Sender};
 
+use crate::parser::aptnode::APTNode;
 use crate::parser::token::Token;
+use crate::pic::coordinatesystem::CoordinateSystem;
+use crate::pic::data::gradient::GradientData;
+use crate::pic::data::grayscale::GrayscaleData;
+use crate::pic::data::hsv::HSVData;
+use crate::pic::data::mono::MonoData;
+use crate::pic::data::rgb::RGBData;
+use crate::pic::pic::Pic;
+
+use ggez::graphics::Color;
 
 // Function pointer definition must be wrapped in a struct to be recursive
 struct StateFunction(fn(&mut Lexer) -> Option<StateFunction>);
@@ -143,6 +153,247 @@ impl<'a> Lexer<'a> {
 
     fn is_linebreak(c: char) -> bool {
         c == '\n'
+    }
+}
+
+pub fn lisp_to_pic(code: String, coord: CoordinateSystem) -> Result<Pic, String> {
+    let mut pic_opt = None;
+    rayon::scope(|s| {
+        let (sender, receiver) = channel();
+        s.spawn(|_| {
+            Lexer::begin_lexing(&code, sender);
+        });
+
+        // TODO: fix race condition that crashes at parser.rs:68. Workaround:
+        std::thread::sleep(std::time::Duration::from_millis(1));
+
+        pic_opt = Some(parse_pic(&receiver, coord))
+    });
+    pic_opt.unwrap()
+}
+
+#[must_use]
+pub fn expect_open_paren(receiver: &Receiver<Token>) -> Result<(), String> {
+    let open_paren = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    match open_paren {
+        Token::OpenParen(_) => Ok(()),
+        Token::Operation(v, line) | Token::Constant(v, line) => {
+            return Err(format!("Expected '(' on line {}, got a '{}'", line, v))
+        }
+        _ => {
+            return Err(format!(
+                "Expected '(' on line {}",
+                extract_line_number(&open_paren)
+            ))
+        }
+    }
+}
+
+#[must_use]
+pub fn expect_close_paren(receiver: &Receiver<Token>) -> Result<(), String> {
+    let close_paren = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    match close_paren {
+        Token::CloseParen(_) => Ok(()),
+        _ => {
+            return Err(format!(
+                "Expected '(' on line {}",
+                extract_line_number(&close_paren)
+            ))
+        }
+    }
+}
+
+#[must_use]
+pub fn expect_operation(s: &str, receiver: &Receiver<Token>) -> Result<(), String> {
+    let op = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    match op {
+        Token::Operation(op_str, _) => {
+            if op_str.to_lowercase() == s {
+                Ok(())
+            } else {
+                Err(format!(
+                    "Expected '{}' on line {}, found {}",
+                    s,
+                    extract_line_number(&op),
+                    op_str
+                ))
+            }
+        }
+        _ => {
+            return Err(format!(
+                "Expected '{}' on line {}, found {:?}",
+                s,
+                extract_line_number(&op),
+                op
+            ))
+        }
+    }
+}
+
+#[must_use]
+pub fn expect_operations(ops: Vec<&str>, receiver: &Receiver<Token>) -> Result<String, String> {
+    let op = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    for s in ops {
+        match op {
+            Token::Operation(op_str, _) => {
+                if op_str.to_lowercase() == s.to_lowercase() {
+                    return Ok(op_str.to_string());
+                }
+            }
+            _ => (),
+        }
+    }
+    return Err(format!(
+        "Unexpected token on line {}",
+        extract_line_number(&op),
+    ));
+}
+
+#[must_use]
+pub fn expect_constant(receiver: &Receiver<Token>) -> Result<f32, String> {
+    let op = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    match op {
+        Token::Constant(vstr, line_number) => {
+            let v = vstr
+                .parse::<f32>()
+                .map_err(|_| format!("Unable to parse number {} on line {}", vstr, line_number))?;
+            Ok(v)
+        }
+        _ => {
+            return Err(format!(
+                "Expected constant on line {}, found {:?}",
+                extract_line_number(&op),
+                op
+            ))
+        }
+    }
+}
+
+pub fn parse_pic(
+    receiver: &Receiver<Token>,
+    coord_default: CoordinateSystem,
+) -> Result<Pic, String> {
+    let mut coord = coord_default;
+    expect_open_paren(receiver)?;
+    let pic_type = receiver.recv().map_err(|_| "Unexpected end of file")?;
+    match pic_type {
+        Token::Operation(s, line_number) => match &s.to_lowercase()[..] {
+            "mono" => {
+                if let Ok(coord_system) = expect_operations(
+                    CoordinateSystem::list_all()
+                        .iter()
+                        .map(|x| x.as_str())
+                        .collect(),
+                    receiver,
+                ) {
+                    coord = coord_system.parse().unwrap();
+                };
+                Ok(Pic::Mono(MonoData {
+                    c: APTNode::parse_apt_node(receiver)?,
+                    coord,
+                }))
+            }
+            "grayscale" => {
+                if let Ok(coord_system) = expect_operations(
+                    CoordinateSystem::list_all()
+                        .iter()
+                        .map(|x| x.as_str())
+                        .collect(),
+                    receiver,
+                ) {
+                    coord = coord_system.parse().unwrap();
+                };
+                Ok(Pic::Grayscale(GrayscaleData {
+                    c: APTNode::parse_apt_node(receiver)?,
+                    coord,
+                }))
+            }
+            "rgb" => {
+                if let Ok(coord_system) = expect_operations(
+                    CoordinateSystem::list_all()
+                        .iter()
+                        .map(|x| x.as_str())
+                        .collect(),
+                    receiver,
+                ) {
+                    coord = coord_system.parse().unwrap();
+                };
+                Ok(Pic::RGB(RGBData {
+                    r: APTNode::parse_apt_node(receiver)?,
+                    g: APTNode::parse_apt_node(receiver)?,
+                    b: APTNode::parse_apt_node(receiver)?,
+                    coord,
+                }))
+            }
+            "hsv" => {
+                if let Ok(coord_system) = expect_operations(
+                    CoordinateSystem::list_all()
+                        .iter()
+                        .map(|x| x.as_str())
+                        .collect(),
+                    receiver,
+                ) {
+                    coord = coord_system.parse().unwrap();
+                };
+                Ok(Pic::HSV(HSVData {
+                    h: APTNode::parse_apt_node(receiver)?,
+                    s: APTNode::parse_apt_node(receiver)?,
+                    v: APTNode::parse_apt_node(receiver)?,
+                    coord,
+                }))
+            }
+            "gradient" => {
+                if let Ok(coord_system) = expect_operations(
+                    CoordinateSystem::list_all()
+                        .iter()
+                        .map(|x| x.as_str())
+                        .collect(),
+                    receiver,
+                ) {
+                    coord = coord_system.parse().unwrap();
+                };
+                let mut colors = Vec::new();
+                expect_open_paren(receiver)?;
+                expect_operation("colors", receiver)?;
+                loop {
+                    let _token = receiver.recv().map_err(|_| "Unexpected end of file")?;
+                    match expect_operations(vec!["color", "stopcolor"], receiver) {
+                        Err(e) => {
+                            if e.starts_with("Unexpected token on line ") {
+                                break;
+                            } else {
+                                panic!("{:?}", e);
+                            }
+                        }
+                        Ok(color_type) => {
+                            let r = expect_constant(receiver)?;
+                            let g = expect_constant(receiver)?;
+                            let b = expect_constant(receiver)?;
+                            if color_type == "color" {
+                                colors.push((Color::new(r, g, b, 1.0), false));
+                            } else {
+                                colors.push((Color::new(r, g, b, 1.0), true));
+                            }
+                            expect_close_paren(receiver)?;
+                        }
+                    }
+                }
+                Ok(Pic::Gradient(GradientData {
+                    colors: colors,
+                    index: APTNode::parse_apt_node(receiver)?,
+                    coord,
+                }))
+            }
+            _ => Err(format!("Unknown pic type {} at line {}", s, line_number)),
+        },
+        _ => Err(format!("Invalid picture type")), //todo line number etc
+    }
+}
+
+pub fn extract_line_number(token: &Token) -> usize {
+    match token {
+        Token::OpenParen(ln) | Token::CloseParen(ln) => *ln,
+        Token::Constant(_, ln) | Token::Operation(_, ln) => *ln,
     }
 }
 
@@ -302,5 +553,14 @@ mod tests {
         assert_eq!(Lexer::is_linebreak('\t'), false);
         assert_eq!(Lexer::is_linebreak('\n'), true);
         assert_eq!(Lexer::is_linebreak('\r'), false);
+    }
+
+    // todo: refactor into a separate module e.g. parser::token
+    #[test]
+    fn test_extract_line_number() {
+        assert_eq!(extract_line_number(&Token::OpenParen(6)), 6);
+        assert_eq!(extract_line_number(&Token::CloseParen(6)), 6);
+        assert_eq!(extract_line_number(&Token::Operation("blablabla", 6)), 6);
+        assert_eq!(extract_line_number(&Token::Constant("blablabla", 6)), 6);
     }
 }
